Add tests for disabling collections in progress and skipping collections on restart.
