+++ tests/fixtures/integration_fixtures.py
@@ -149,15 +149,17 @@ def postgres_inserts(postgres_integration_db):
 # ======================= mongodb  ==========================
 
 
-@pytest.fixture(scope="session")
-def integration_mongodb_config() -> ConnectionConfig:
-    return ConnectionConfig(
+@pytest.fixture(scope="function")
+def integration_mongodb_config(db) -> ConnectionConfig:
+    connection_config = ConnectionConfig(
         key="mongo_example",
         connection_type=ConnectionType.mongodb,
         access=AccessLevel.write,
         secrets=integration_secrets["mongo_example"],
         name="mongo_example",
     )
+    yield connection_config
+    connection_config.delete(db)
 
 
 @pytest.fixture(scope="session")
